test: add coverage for stats models
